Day 5 part 2 continued: I can also do this the "right" way....
Note that the power set of pairs of letters is much too large a state
space to use with Tabulate.  (Yes, yes, we're not using the *entire*
power set, mathematically speaking, but I'm pretty sure this covers
enough to be wildly intractable.)  So this is a little disappointing, but
it'll be interesting to benchmark this against `grep` backreferences.