Day 6: Traits FTW.
Sadly, my fancy front-to-back visible-area traversal thing is not hugely
useful in nordic mode, because basically nothing can be pruned.  It would
be possible to adapt it (conjecture: any string of off/on/toggle is
equivalent to n*off + m*on for some n,m>0) but I'm not sure I care.