Day 7: Higher-ranked trait bounds exist and can actually be used here.
This lets me get rid of the macro sadness in `shells.rs` and just use
the type system.  Yay.  (Also means creating another family of size-0
types to host a type-parameterized factory pattern, but that's not *so*
bad.)

The first version (Fri Dec 11 02:03:44 2015 -0500) would have made a
mess of `lazy.rs` without converting it into the current fake-HKT style
first, so I commented it all out and temporarily changed `main.rs` to use
`eager`... which causes exponential time blowup for the problem inputs.
Which is why it's squashed with the followup instead of separate.